nixos: unify virtual console options
This commit moves all the virtual console related options to a dedicated
config/console.nix NixOS module.

Currently most of these are defined in config/i18n.nix with a "console" prefix
like `i18n.consoleFont`, `i18n.consoleColors` or under `boot` and are
implemented in tasks/kbd.nix.

Since they have little to do with actual internationalisation and
are (informally) in an attrset already, it makes sense to move them to a
specific module.